move data to separate repo, update mgmt commands and readme
