+++ rorapi/management/commands/setup.py
@@ -1,23 +1,34 @@
+import requests
 import zipfile
 
 from django.core.management.base import BaseCommand
 from .deleteindex import Command as DeleteIndexCommand
 from .createindex import Command as CreateIndexCommand
-from .indexgrid import Command as IndexGridCommand
+from .indexrordump import Command as IndexRorDumpCommand
 from rorapi.settings import GRID, ROR_DUMP
 
 
 class Command(BaseCommand):
     help = 'Setup ROR API'
 
+    def add_arguments(self, parser):
+        parser.add_argument('filename', type=str, help='Name of data dump zip file to index without extension')
+
     def handle(self, *args, **options):
-        # make sure ROR JSON file exists
-        if not zipfile.is_zipfile(ROR_DUMP['ROR_ZIP_PATH']):
-            self.stdout.write('ROR dataset for GRID version {} not found. '.
-                              format(GRID['VERSION']) +
-                              'Please run the upgrade command first.')
+        # make sure ROR dump file exists
+        filename = options['filename']
+        url = ROR_DUMP['URL'] + filename + '.zip'
+        print(url)
+        ror_dump_zip = requests.get(url)
+
+        if ror_dump_zip.status_code == 200:
+            DeleteIndexCommand().handle(*args, **options)
+            CreateIndexCommand().handle(*args, **options)
+            IndexRorDumpCommand().handle(*args, **options)
+
+        else:
+            self.stdout.write('ROR dataset for version {} not found. '.
+                              format(filename) +
+                              'Please generate the data dump first.')
             return
 
-        DeleteIndexCommand().handle(args, options)
-        CreateIndexCommand().handle(args, options)
-        IndexGridCommand().handle(args, options)
